Detection of Echelon Cards Use only if detected Add PLAYBACK card (untested) Add MENU card (working) Add PROGRAM card (working) USB-HID Keymap to PROGRAM card. sbuttons now 16bit value.

+++ lib/cardbus/surface_buttons.h
@@ -9,7 +9,7 @@
 
 // updated every loop, register of each surface button state
 // of buttons defined in button_def.h
-static uint8_t sbuttons[TOTAL_BUTTONS];
+static uint16_t sbuttons[TOTAL_BUTTONS];
 
 #define SHIFT_HELD  (sbuttons[BTN_SHIFT]!=0)
 
@@ -81,7 +81,7 @@ class SKeyboard {
 
 
 void SKeyboard::update(){
-    for (uint8_t cnt = 0 ; cnt < TOTAL_BUTTONS ; cnt++) {
+    for (uint16_t cnt = 0 ; cnt < TOTAL_BUTTONS ; cnt++) {
       
       if ((laststate[cnt] != 0) & (sbuttons[cnt] == 0)) {
        keyQueue.push(cnt);
